requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.5
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
wincertstore==0.2
